(PC-23190)[API] chore: extract patch price category test
